Added segmentation window and function (#11)
* Seperated external windows into new file, added load metadata function when 'Load Metadata' is clicked. For now, it only prints out the path to the selected text file

* Added segmentation window and behaviour
